Fixed the bug where arrays would not save after being saved.
Arrays were not saving after being loaded and unloaded twice this was due to a bug of the value not being stored in the refelected feild. 
Reflection helper has not new helper functions. 
Nested types that have parent that are null no longer try to create there parent but just throw an exception. Might add a function later to do that. 
Added a new function to log all of the fields and their content.