Major Refactor of the core idea
Before everything was combined into one class which made it super hard to debug and was not very clear how it worked. I took the design from TinyJson and it's proxy object system and applied it to this. Currently still working out some bugs but it's finally in working order.