Changed how objects are saved and loaded
Before each class would have a property path which was the path to the root seprated with spaces. It would then use the reflection helper to find it's value from the root. This caused big problems when dealing with value types as they would get erased. Now each object get set a refernece from it's parent to save and read. This fixes all my issues but requires some reworking. Arrays and Dictionaries are currently commented out while I update them.